tv_titles: remove dependency on titlecase
It's unnecessary, and only provides a minor convenience. Not sure if all
Wikipedia titles are titlecase'd the same way as this function, so
instead of refactoring to check whether we 404 and use the raw input,
we'll just remove the convenience. Development seems to have stalled,
and the AUR package is orphaned.