tesla dashcam viewer: Use mpv, fix playlist order
Play videos using mpv player instead of VLC. In addition, re-order the
videos for playback.

Previously, we'd move the video containing the event to the front of the
playlist queue so vlc would always load that first. Maybe VLC didn't
have a command-line option to jump to an index in the play queue?